pyproject.toml
[tool.poetry]
name = "local-server"
version = "0.1.0"
description = "Maybe Websocket is a great protocol."
authors = ["cloudw233 <qrz20081004@outlook.com>"]
license = "AGPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
pydantic = "^2.10.6"
httpx = "^0.28.1"
fastapi = "^0.115.10"
tortoise-orm = "^0.24.1"
openai = "^1.65.1"
loguru = "^0.7.3"
tomlkit = "^0.13.2"
orjson = "^3.10.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
